web_interface/requirements.txt
flask
gunicorn
psycopg2-binary
pyyaml
numpy
